+++ demo/CypressLayout.tsx
@@ -22,6 +22,7 @@ import {
   FileDrop,
   FileUpload,
   Input,
+  LinkTree,
   Loader,
   MaskedInput,
   Menu,
@@ -115,6 +116,9 @@ export const CypressLayout = () => (
         </Route>
         <Route path="/cypress/input">
           <Input />
+        </Route>
+        <Route path="/cypress/link-tree">
+          <LinkTree />
         </Route>
         <Route path="/cypress/loader">
           <Loader />