+++ demo/components/LinkTree/index.tsx
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { Story } from '../Story';
+import { Minimal } from './Minimal';
+
+export const LinkTree = (): React.ReactElement => (
+  <Story title="LinkTree">
+    <Minimal title="Minimal example" />
+  </Story>
+);