+++ demo/cypress/index.ts
@@ -17,6 +17,7 @@ export { DropZone } from './DropZone';
 export { FileDrop } from './FileDrop';
 export { FileUpload } from './FileUpload';
 export { Input } from './Input';
+export { LinkTree } from './LinkTree';
 export { Loader } from './Loader';
 export { MaskedInput } from './MaskedInput';
 export { Menu } from './Menu';