+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.43.0] - 2021-03-29
 
 ### Added
+- LinkTree: Added new component
 - Drawer: Added new component
 
 ### Fixed